Remove KeyValueDb now that it's not used directly (merge into KeyValueJsonDb)
